Feat: Move function to new module group. #103
